Day7/Gen app1/requirements.txt
langchain
langchain_community
langchain_groq
streamlit
unstructured
load_dotenv
